requirements.txt
django
djangorestframework
crewai
python-dotenv
django-cors-headers
litellm==1.33.0